pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["daniskzn <danis.sharafiev@bk.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
ruff = "^0.3.2"
drf-spectacular = "^0.27.1"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.1"
requests = "^2.31.0"
django-debug-toolbar = "^4.3.0"
pika = "^1.3.2"
faststream = {extras = ["rabbit"], version = "^0.4.7"}
uvicorn = {extras = ["standard"], version = "^0.28.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
